refactor : Post 삭제된 댓글 재삭제시 삭제시간 업데이트 되던 오류 수정

+++ src/main/java/com/missionx/questloggers/domain/post/service/PostService.java
@@ -2,6 +2,7 @@ package com.missionx.questloggers.domain.post.service;
 
 import com.missionx.questloggers.domain.post.dto.*;
 import com.missionx.questloggers.domain.post.entity.Post;
+import com.missionx.questloggers.domain.post.exception.AlreadyDeletedPostException;
 import com.missionx.questloggers.domain.post.exception.NotFoundPostException;
 import com.missionx.questloggers.domain.post.repository.PostRepository;
 import com.missionx.questloggers.domain.user.entity.User;
@@ -40,7 +41,7 @@ public class PostService {
     @Transactional
     public UpdatePostResponseDto updatePostService(Long postId, UpdatePostRequestDto updatePostRequestDto) {
         Post foundPost = postRepository.findById(postId)
-                .orElseThrow(() -> new NotFoundPostException(HttpStatus.NOT_FOUND, "post not found"));
+                .orElseThrow(() -> new NotFoundPostException(HttpStatus.NOT_FOUND, "게시글을 찾을 수 없습니다."));
         foundPost.updatePost(updatePostRequestDto);
         return new UpdatePostResponseDto(foundPost.getId(), foundPost.getTitle(), foundPost.getContent());
     }
@@ -67,7 +68,7 @@ public class PostService {
     @Transactional(readOnly = true)
     public GetPostResponseDto getPostService(Long postId) {
         Post foundPost = postRepository.findById(postId)
-                .orElseThrow(()-> new RuntimeException("post not found"));
+                .orElseThrow(()-> new RuntimeException("게시글을 찾을 수 없습니다."));
 
         return new GetPostResponseDto(foundPost.getUser().getId(), foundPost.getId(), foundPost.getTitle(), foundPost.getContent());
     }
@@ -75,8 +76,13 @@ public class PostService {
     @Transactional
     public void deletePostService(Long postId) {
         Post foundPost = postRepository.findById(postId)
-                .orElseThrow(() -> new RuntimeException("post not found"));
-        foundPost.delete();
+                .orElseThrow(() -> new RuntimeException("게시글을 찾을 수 없습니다."));
+        if (foundPost.getDeletedAt() == null) {
+            foundPost.delete();
+        } else {
+            throw new AlreadyDeletedPostException(HttpStatus.NOT_FOUND, "이미 삭제된 게시글입니다.");
+        }
+
     }
 
     // 다른 domain에서 사용하는 기능
